In-memory engine: implement core local array and collect read flow statistics (#16945)
ref tikv/tikv#16141

implement core local array and collect read flow statistics

+++ components/engine_traits/src/iterable.rs
@@ -118,9 +118,9 @@ pub trait RefIterable {
 }
 
 pub trait IterMetricsCollector {
-    fn internal_delete_skipped_count(&self) -> usize;
+    fn internal_delete_skipped_count(&self) -> u64;
 
-    fn internal_key_skipped_count(&self) -> usize;
+    fn internal_key_skipped_count(&self) -> u64;
 
     // todo: add more metrics related methods when needed.
 }
